Add type cast to ensure max_attempts config returns integer value

+++ src/Framework/Uploads/TransformJob.php
@@ -34,7 +34,7 @@ class TransformJob extends Job
      */
     public function maxAttempts(): int
     {
-        return Container::getInstance()
+        return (int) Container::getInstance()
             ->get('config')
             ->get('uploads.max_attempts', 1);
     }
